+++ server/src/api/webex.ts
@@ -69,6 +69,8 @@ async function updateAccount(req: Request, res: Response, next: NextFunction) {
 	const user = req.user;
 	const groupId = req.group!.id;
 	const accountId = Number(req.params.accountId);
+	if (isNaN(accountId))
+		return res.status(404).send("path parameter accountId not a number");
 	try {
 		const changes = webexAccountChangeSchema.parse(req.body);
 		const data = await updateWebexAccount(
@@ -92,6 +94,8 @@ async function revokeAccountAuth(
 	const user = req.user;
 	const groupId = req.group!.id;
 	const accountId = Number(req.params.accountId);
+	if (isNaN(accountId))
+		return res.status(404).send("path parameter accountId not a number");
 	try {
 		const data = await revokeAuthWebexAccount(
 			req,
@@ -108,6 +112,8 @@ async function revokeAccountAuth(
 async function removeAccount(req: Request, res: Response, next: NextFunction) {
 	const groupId = req.group!.id;
 	const accountId = Number(req.params.accountId);
+	if (isNaN(accountId))
+		return res.status(404).send("path parameter accountId not a number");
 	try {
 		const data = await deleteWebexAccount(groupId, accountId);
 		res.json(data);
@@ -160,14 +166,14 @@ async function removeMeetings(req: Request, res: Response, next: NextFunction) {
 
 const router = Router();
 router
-	.all("*", validatePermissions)
+	.all(/(.*)/, validatePermissions)
 	.route("/accounts")
 	.get(getAccounts)
 	.post(addAccount);
 router
-	.patch("/accounts/:accountId(\\d+)", updateAccount)
-	.patch("/accounts/:accountId(\\d+)/revoke", revokeAccountAuth)
-	.delete("/accounts/:accountId(\\d+)", removeAccount);
+	.patch("/accounts/:accountId", updateAccount)
+	.patch("/accounts/:accountId/revoke", revokeAccountAuth)
+	.delete("/accounts/:accountId", removeAccount);
 router
 	.route("/meetings")
 	.get(getMeetings)
